Refactor instrument module into separate instrument and camera modules
